+++ cli/cli/commands/enclave/rm/rm.go
@@ -6,6 +6,9 @@ import (
 	"fmt"
 	"github.com/docker/docker/client"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/args"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/flags"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
@@ -13,32 +16,58 @@ import (
 	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
-	"github.com/spf13/cobra"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"strings"
 )
 
 const (
-	shouldForceRemoveArg = "force"
-	enclaveIdArg        = "enclave-id"
+	shouldForceRemoveFlagKey = "force"
+	enclaveIdArgKey          = "enclave-id"
 
 	defaultShouldForceRemove = false
 )
 
+var EnclaveRmCmd = &kurtosis_command.KurtosisCommand{
+	CommandStr:       command_str_consts.EnclaveRmCmdStr,
+	ShortDescription: "Destroys the specified enclaves",
+	LongDescription:  "Destroys the specified enclaves, removing all resources associated with them",
+	Flags:            []*flags.FlagConfig{
+		{
+			Key:       shouldForceRemoveFlagKey,
+			Usage:     "Deletes all enclaves, regardless of whether they're already stopped",
+			Shorthand: "f",
+			Type:      flags.FlagType_Bool,
+			Default:   "false",
+		},
+	},
+	// TODO Use a prebuilt enclaveIdArg component here!!!
+	Args:             []*args.ArgConfig{
+		{
+			Key:             enclaveIdArgKey,
+			IsGreedy:        true,
+			CompletionsFunc: nil,
+			ValidationFunc:  nil,
+		},
+	},
+	RunFunc:          run,
+}
+
+/*
 var RmCmd = &cobra.Command{
-	Use:   command_str_consts.EnclaveRmCmdStr + " [flags] " + enclaveIdArg + " [" + enclaveIdArg + "...]",
+	Use:   command_str_consts.EnclaveRmCmdStr + " [flags] " + enclaveIdArgKey + " [" + enclaveIdArgKey + "...]",
 	DisableFlagsInUseLine: true,
 	Short: "Destroys the specified enclaves",
 	Long: "Destroys the specified enclaves, removing all resources associated with them",
 	RunE:  run,
 }
 
+
 var shouldForceRemove bool
 
 func init() {
 	RmCmd.Flags().BoolVarP(
 		&shouldForceRemove,
-		shouldForceRemoveArg,
+		shouldForceRemoveFlagKey,
 		"f",
 		defaultShouldForceRemove,
 		"Deletes all enclaves, regardless of whether they're already stopped",
@@ -47,14 +76,20 @@ func init() {
 }
 
 func run(cmd *cobra.Command, args []string) error {
+*/
+func run(flags *flags.ParsedFlags, args *args.ParsedArgs) error {
 	ctx := context.Background()
 
-	if len(args) == 0 {
-		return stacktrace.NewError("At least one enclave ID to remove must be provided")
+	allEnclaveIds, err := args.GetGreedyArg(enclaveIdArgKey)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting the enclave IDs using arg key '%v'; this is a bug in Kurtosis!", enclaveIdArgKey)
+	}
+	shouldForceRemove, err := flags.GetBool(shouldForceRemoveFlagKey)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting the force-removal flag value using key '%v'; this is a bug in Kurtosis!", shouldForceRemoveFlagKey)
 	}
 
 	logrus.Info("Destroying enclaves...")
-
 	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating the Docker client")
@@ -78,8 +113,8 @@ func run(cmd *cobra.Command, args []string) error {
 	allEnclaveInfo := getEnclavesResp.EnclaveInfo
 
 	enclaveDestructionErrorStrs := []string{}
-	for _, enclaveId := range args {
-		if err := destroyEnclave(ctx, enclaveId, allEnclaveInfo, engineClient); err != nil {
+	for _, enclaveId := range allEnclaveIds {
+		if err := destroyEnclave(ctx, enclaveId, allEnclaveInfo, engineClient, shouldForceRemove); err != nil {
 			enclaveDestructionErrorStrs = append(enclaveDestructionErrorStrs, err.Error())
 		}
 	}
@@ -100,7 +135,13 @@ func run(cmd *cobra.Command, args []string) error {
 // ====================================================================================================
 // 									   Private helper methods
 // ====================================================================================================
-func destroyEnclave(ctx context.Context, enclaveId string, allEnclaveInfo map[string]*kurtosis_engine_rpc_api_bindings.EnclaveInfo, engineClient kurtosis_engine_rpc_api_bindings.EngineServiceClient) error {
+func destroyEnclave(
+	ctx context.Context,
+	enclaveId string,
+	allEnclaveInfo map[string]*kurtosis_engine_rpc_api_bindings.EnclaveInfo,
+	engineClient kurtosis_engine_rpc_api_bindings.EngineServiceClient,
+	shouldForceRemove bool,
+) error {
 	enclaveInfo, found := allEnclaveInfo[enclaveId]
 	if !found {
 		return stacktrace.NewError("No enclave '%v' exists", enclaveId)
@@ -122,7 +163,7 @@ func destroyEnclave(ctx context.Context, enclaveId string, allEnclaveInfo map[st
 			"Refusing to destroy enclave '%v' because its status is '%v'; to force its removal, rerun this command with the '%v' flag",
 			enclaveId,
 			enclaveStatus,
-			shouldForceRemoveArg,
+			shouldForceRemoveFlagKey,
 		)
 	}